work on LocalBGS specs
