+++ modules/claims_api/spec/lib/claims_api/local_bgs_spec.rb
@@ -7,6 +7,14 @@ require 'claims_api/error/soap_error_handler'
 describe ClaimsApi::LocalBGS do
   subject { described_class.new external_uid: 'xUid', external_key: 'xKey' }
 
+  before do
+    allow(Flipper).to(
+      receive(:enabled?)
+        .with(:claims_api_local_bgs_refactor)
+        .and_return(false)
+    )
+  end
+
   let(:soap_error_handler) { ClaimsApi::SoapErrorHandler.new }
 
   describe '#find_poa_by_participant_id' do
@@ -66,7 +74,9 @@ describe ClaimsApi::LocalBGS do
       it 'returns an empty array' do
         expect(error_message.count).to eq(2) # trick the claims count check
         # error message should trigger return
-        allow(subject_instance).to receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(error_message)
+        allow(subject_instance.proxied).to(
+          receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(error_message)
+        )
         expect(subject.all(id)).to eq([]) # verify correct return
       end
     end
@@ -76,7 +86,9 @@ describe ClaimsApi::LocalBGS do
         VCR.use_cassette('claims_api/bgs/claims/claims_trimmed_down') do
           claims = subject_instance.find_benefit_claims_status_by_ptcpnt_id('600061742')
           claims[:benefit_claims_dto][:benefit_claim] = claims[:benefit_claims_dto][:benefit_claim][0]
-          allow(subject_instance).to receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(claims)
+          allow(subject_instance.proxied).to(
+            receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(claims)
+          )
 
           begin
             ret = subject_instance.send(:transform_bgs_claims_to_evss, claims)
@@ -93,7 +105,9 @@ describe ClaimsApi::LocalBGS do
     context 'when an empty array gets returned it still does not pass the count check' do
       it 'returns an empty array' do
         # error message should trigger return
-        allow(subject_instance).to receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(empty_array)
+        allow(subject_instance.proxied).to(
+          receive(:find_benefit_claims_status_by_ptcpnt_id).with(id).and_return(empty_array)
+        )
         expect(subject.all(id)).to eq([]) # verify correct return
       end
     end
